% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{ddg.clear.detail}
\alias{ddg.clear.detail}
\title{ddg.clear.detail clears the current value of provenance detail.}
\usage{
ddg.clear.detail()
}
\value{
nothing
}
\description{
ddg.clear.detail clears the current value of provenance detail.
}
