% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Console.R
\name{ddg.console.off}
\alias{ddg.console.off}
\title{ddg.console.off creates a final console node and then
turns off the console mode.  If we are not in console mode,
it does nothing.}
\usage{
ddg.console.off()
}
\value{
nothing
}
\description{
ddg.console.off creates a final console node and then
turns off the console mode.  If we are not in console mode,
it does nothing.
}
