% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Console.R
\name{ddg.console.on}
\alias{ddg.console.on}
\title{ddg.console.on turns on the console mode of DDG construction.
If it was not already on, it writes a timestamp to the history
file.}
\usage{
ddg.console.on()
}
\value{
nothing
}
\description{
ddg.console.on turns on the console mode of DDG construction.
If it was not already on, it writes a timestamp to the history
file.
}
