% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{ddg.details.omitted}
\alias{ddg.details.omitted}
\title{ddg.details.omitted inserts an operational node called "Details Omitted"
in cases where not all iterations of a loop are annotated.  This may
happen if the number of the first loop to be annotaed (first.loop) is
greater than 1 and/or if the total number of loops to be annotated is
less than the actual number of iterations.
It also sets a variable to remember that the last construct is incomplete
so that the right data nodes get created.
NOTE:  This might be useful outside of the context of loops, but is
currently only used within loops (internal use only)}
\usage{
ddg.details.omitted()
}
\value{
nothing
}
\description{
ddg.details.omitted inserts an operational node called "Details Omitted"
in cases where not all iterations of a loop are annotated.  This may
happen if the number of the first loop to be annotaed (first.loop) is
greater than 1 and/or if the total number of loops to be annotated is
less than the actual number of iterations.
It also sets a variable to remember that the last construct is incomplete
so that the right data nodes get created.
NOTE:  This might be useful outside of the context of loops, but is
currently only used within loops (internal use only)
}
