% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{ddg.display}
\alias{ddg.display}
\title{ddg.display loads & displays the current provenance graph in DDG Explorer}
\usage{
ddg.display()
}
\value{
nothing
}
\description{
ddg.display loads & displays the current provenance graph in DDG Explorer
}
