% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{ddg.eval}
\alias{ddg.eval}
\title{ddg.eval evaluates a statement and creates data flow edges from
variable and function return nodes that are used in the
statement. If the statement is an assignment statement, it also
creates a data node for the variable assigned and a corresponding
data flow edge. If ddg.eval is called from inside a function, cmd.func
is a function that returns the corresponding DDGStatement object.
If ddg.eval is called from inside a control block, cmd.func is an
integer that points to the corresponding DDGStatement object stored
in the list .ddg.statements (internal use only)}
\usage{
ddg.eval(statement, cmd.func = NULL)
}
\arguments{
\item{statement}{a string version of the statement to evaluate.}

\item{cmd.func}{the corresponding DDGStatement if inside a function,
or an integer identifying the position of the statement in a list
if inside a control construct}
}
\description{
ddg.eval evaluates a statement and creates data flow edges from
variable and function return nodes that are used in the
statement. If the statement is an assignment statement, it also
creates a data node for the variable assigned and a corresponding
data flow edge. If ddg.eval is called from inside a function, cmd.func
is a function that returns the corresponding DDGStatement object.
If ddg.eval is called from inside a control block, cmd.func is an
integer that points to the corresponding DDGStatement object stored
in the list .ddg.statements (internal use only)
}
