% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{ddg.finish}
\alias{ddg.finish}
\title{ddg.finish creates a procedure node of type Finish called pname.
Users can right-click on a finish node in DDG Explorer and see
the code between start and finish nodes in the original script.}
\usage{
ddg.finish(pname = NULL)
}
\arguments{
\item{pname}{the label for the node. This can be passed as
a string or as a name. It can be omitted if ddg.finish is called
by a function, in which case the name of the function will be used.}
}
\description{
ddg.finish creates a procedure node of type Finish called pname.
Users can right-click on a finish node in DDG Explorer and see
the code between start and finish nodes in the original script.
}
