% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops.R
\name{ddg.first.loop}
\alias{ddg.first.loop}
\title{ddg.first.loop returns the loop iteration at which we should start
collecting provenance (internal use only)}
\usage{
ddg.first.loop()
}
\value{
the first iteration to collect provenance
}
\description{
ddg.first.loop returns the loop iteration at which we should start
collecting provenance (internal use only)
}
