% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops.R
\name{ddg.forloop}
\alias{ddg.forloop}
\title{ddg.forloop inserts a procedure node and a data node in a for loop,
indicating the value currently assigned to the index variable
(internal use only)}
\usage{
ddg.forloop(index.var)
}
\arguments{
\item{index.var}{a parsed expression containing the index variable}
}
\value{
nothing
}
\description{
ddg.forloop inserts a procedure node and a data node in a for loop,
indicating the value currently assigned to the index variable
(internal use only)
}
