% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{ddg.function}
\alias{ddg.function}
\title{ddg.function creates a procedure node of type Operation for
procedures implemented as functions in the original R script.
The function name and input parameters are obtained automatically
from the calling environment. The outs parameters may be used
optionally to create output data nodes. These MUST be passed as
a list of strings, not names, unless the value is a file name.
Users can right-click on the procedure node in DDG Explorer to
see the code for the function in the original script. For more
details on outs parameters, see .ddg.create.output.nodes.
The outs parameters should be a list of names of data nodes to be created as
outputs to this procedure node. These MUST be passed as
a list of strings, not names, unless the value is a file name
(internal use only)}
\usage{
ddg.function(outs.graphic = NULL, outs.data = NULL, outs.exception = NULL,
  outs.url = NULL, outs.file = NULL, graphic.fext = "jpeg")
}
\arguments{
\item{outs.graphic}{(optional) the name of a snapshot node to be used as a
file name.  A graphical snapshot is simply a captured image
of the graphic device active at the time of the call to
ddg.function or ddg.procedure.}

\item{outs.data}{(optional) a list of names of data nodes.}

\item{outs.exception}{(optional) a list of names of exception nodes.}

\item{outs.url}{(optional) a list of names of url nodes.}

\item{outs.file}{(optional) a list of names of file nodes. Supported file
extensions include: .csv, .jpg, .jpeg, .pdf, and .txt.}

\item{graphic.fext}{(optional) the file extension for a graphics file, defaults to jpeg.}
}
\value{
nothing
}
\description{
ddg.function creates a procedure node of type Operation for
procedures implemented as functions in the original R script.
The function name and input parameters are obtained automatically
from the calling environment. The outs parameters may be used
optionally to create output data nodes. These MUST be passed as
a list of strings, not names, unless the value is a file name.
Users can right-click on the procedure node in DDG Explorer to
see the code for the function in the original script. For more
details on outs parameters, see .ddg.create.output.nodes.
The outs parameters should be a list of names of data nodes to be created as
outputs to this procedure node. These MUST be passed as
a list of strings, not names, unless the value is a file name
(internal use only)
}
