% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{ddg.get.detail}
\alias{ddg.get.detail}
\title{ddg.detail returns the current level of provenance detail.}
\usage{
ddg.get.detail()
}
\value{
the current level of detail (0-3)
}
\description{
ddg.detail returns the current level of provenance detail.
}
