% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{ddg.json}
\alias{ddg.json}
\title{ddg.json returns the current provenance graph as a prov-json string}
\usage{
ddg.json()
}
\value{
the current provenance graph
}
\description{
ddg.json returns the current provenance graph as a prov-json string
}
