% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops.R
\name{ddg.loop.annotate.off}
\alias{ddg.loop.annotate.off}
\title{ddg.loop.annotate.off turns off loop annotation (internal use only)}
\usage{
ddg.loop.annotate.off()
}
\value{
nothing
}
\description{
ddg.loop.annotate.off turns off loop annotation (internal use only)
}
