% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops.R
\name{ddg.loop.annotate.on}
\alias{ddg.loop.annotate.on}
\title{ddg.loop.annotate.on turns on loop annotation (internal use only)}
\usage{
ddg.loop.annotate.on()
}
\value{
nothing
}
\description{
ddg.loop.annotate.on turns on loop annotation (internal use only)
}
