% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops.R
\name{ddg.loop.count.inc}
\alias{ddg.loop.count.inc}
\title{ddg.loop.count.inc increments the current count for the specified loop
(internal use only)}
\usage{
ddg.loop.count.inc(loop.num)
}
\arguments{
\item{loop.num}{the id of the loop being executed}
}
\value{
the updated value of the counter.
}
\description{
ddg.loop.count.inc increments the current count for the specified loop
(internal use only)
}
