% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops.R
\name{ddg.loop.count}
\alias{ddg.loop.count}
\title{ddg.loop.count returns the number of times a loop has iterated so far
(internal use only)}
\usage{
ddg.loop.count(loop.num)
}
\arguments{
\item{loop.num}{the id of the loop to look up}
}
\value{
the number of times the loop has iterated
}
\description{
ddg.loop.count returns the number of times a loop has iterated so far
(internal use only)
}
