% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops.R
\name{ddg.max.loops}
\alias{ddg.max.loops}
\title{ddg.max.loops returns the maximum number of iterations of a loop for
which provenance should be collected}
\usage{
ddg.max.loops()
}
\value{
the maximum number of iterations to collect provenance
}
\description{
ddg.max.loops returns the maximum number of iterations of a loop for
which provenance should be collected
}
