% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataNodes.R
\name{ddg.max.snapshot.size}
\alias{ddg.max.snapshot.size}
\title{ddg.max.snapshot.size returns the maximum size for snapshots in KB}
\usage{
ddg.max.snapshot.size()
}
\value{
maximum snapshot size in KB
}
\description{
ddg.max.snapshot.size returns the maximum size for snapshots in KB
}
