% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops.R
\name{ddg.not.inside.loop}
\alias{ddg.not.inside.loop}
\title{ddg.not.inside.loop decreases the counter of the level of nesting of loops
(internal use only)}
\usage{
ddg.not.inside.loop()
}
\value{
nothing
}
\description{
ddg.not.inside.loop decreases the counter of the level of nesting of loops
(internal use only)
}
