% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{ddg.quit}
\alias{ddg.quit}
\title{ddg.quit saves the current provenance graph and closes the current ddg}
\usage{
ddg.quit(save.debug = FALSE)
}
\arguments{
\item{save.debug}{(optional) - If TRUE, save debug files to debug directory.
Used in console mode.}
}
\value{
nothing
}
\description{
ddg.quit saves the current provenance graph and closes the current ddg
}
