% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops.R
\name{ddg.reset.loop.count}
\alias{ddg.reset.loop.count}
\title{ddg.reset.loop.count sets the current count for the specified loop to zero
(internal use only)}
\usage{
ddg.reset.loop.count(loop.num)
}
\arguments{
\item{loop.num}{the id of the loop to reset}
}
\value{
nothing
}
\description{
ddg.reset.loop.count sets the current count for the specified loop to zero
(internal use only)
}
