% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{ddg.return.value}
\alias{ddg.return.value}
\title{ddg.return.value creates a data node for a function's return value. If
the function is called from a console command and console mode is
enabled, a data flow edge will be created linking this node to
the console command that uses the value. ddg.return.value returns the
same value as the function (expr) and can be used in place of the
function's normal return statement(s) if it is the last statement
in the function.  Otherwise, it should be a parameter to return,
as in return(ddg.return.value(expr)). If expr is an assignment, nodes
and edges are created for the assignment (internal use only)}
\usage{
ddg.return.value(expr = NULL, cmd.func = NULL)
}
\arguments{
\item{expr}{the value returned by the function.}

\item{cmd.func}{the DDGStatement object for the return statement}
}
\value{
the value that the function whose return value we are capturing 
   returns
}
\description{
ddg.return.value creates a data node for a function's return value. If
the function is called from a console command and console mode is
enabled, a data flow edge will be created linking this node to
the console command that uses the value. ddg.return.value returns the
same value as the function (expr) and can be used in place of the
function's normal return statement(s) if it is the last statement
in the function.  Otherwise, it should be a parameter to return,
as in return(ddg.return.value(expr)). If expr is an assignment, nodes
and edges are created for the assignment (internal use only)
}
