% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{ddg.run}
\alias{ddg.run}
\title{ddg.run initiates execution of a script, collecting provenance as it executes}
\usage{
ddg.run(r.script.path = NULL, ddgdir = NULL, overwrite = TRUE, f = NULL,
  annotate.inside.functions = TRUE, first.loop = 1, max.loops = 1,
  max.snapshot.size = 10, save.debug = FALSE, display = FALSE,
  hash.algorithm = "md5")
}
\arguments{
\item{r.script.path}{(optional) - the full path to the R script.
If provided, a copy of the script will be saved with the DDG.
If only r.script.path is provided, the script is sourced using
ddg.source and a DDG is created for the script.}

\item{ddgdir}{(optional) - the directory where the DDG will be saved.
If not provided, the DDG will be saved in a directory inside R's
temporary directory}

\item{overwrite}{(optional) - if TRUE, the ddg is overwritten each time
the script is executed.  If FALSE, a timestamp is attached to the ddg
folder name so that the provenance of earlier runs is not overwritten.}

\item{f}{(optional) - a function to run. If supplied, the function f
is executed with calls to ddg.init and ddg.save so that
provenance for the function is captured.  Exactly one of f and r.script.path
should be provided.}

\item{annotate.inside.functions}{(optional) - if TRUE, functions are annotated.}

\item{first.loop}{(optional) - the first loop to annotate in a for, while, or
repeat statement.}

\item{max.loops}{(optional) - the maximum number of loops to annotate in a for,
while, or repeat statement. If max.loops = -1 there is no limit.
If max.loops = 0, no loops are annotated.  If non-zero, if-statements
are also annotated.}

\item{max.snapshot.size}{(optional) - the maximum size for objects that
should be output to snapshot files. If 0, no snapshot files are
saved. If -1, all snapshot files are saved.  Size in kilobytes.
Note that this tests the size of the object that will be turned
into a snapshot, not the size of the resulting snapshot.}

\item{save.debug}{(optional) - If TRUE, save debug files to debug directory.}

\item{display}{(optional) - if TRUE, display provenance graph in DDG Explorer}

\item{hash.algorithm}{(optional) - If save.hashtable is true, this allows the caller to 
select the hash algorithm to use. This uses the digest function from the digest package.
The choices are md5, which is also the default, sha1, crc32, sha256, sha512, xxhash32, xxhash64 and murmur32.}
}
\value{
nothing
}
\description{
ddg.run initiates execution of a script, collecting provenance as it executes
}
