% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{ddg.set.detail}
\alias{ddg.set.detail}
\title{ddg.set.detail sets the level of provenance detail to be collected.
If ddg.detail is not set, the values of annotate.inside, max.loops,
and max.snapshot.size passed to ddg.run are used instead.
0 = no internal annotation, no snapshots.
1 = 1 loop, snapshots < 10k.
2 = 10 loops, snapshots < 100k.
3 = all loops, all snapshots.}
\usage{
ddg.set.detail(detail.level)
}
\arguments{
\item{detail.level}{level of detail to set (0-3)}
}
\value{
nothing
}
\description{
ddg.set.detail sets the level of provenance detail to be collected.
If ddg.detail is not set, the values of annotate.inside, max.loops,
and max.snapshot.size passed to ddg.run are used instead.
0 = no internal annotation, no snapshots.
1 = 1 loop, snapshots < 10k.
2 = 10 loops, snapshots < 100k.
3 = all loops, all snapshots.
}
