% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{ddg.should.run.annotated}
\alias{ddg.should.run.annotated}
\title{ddg.should.run.annotated returns True if we should run the annotated
version of a function (internal use only)}
\usage{
ddg.should.run.annotated(func.name)
}
\arguments{
\item{func.name}{name of function}
}
\value{
True if we should run annotated version
}
\description{
ddg.should.run.annotated returns True if we should run the annotated
version of a function (internal use only)
}
