% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{ddg.source}
\alias{ddg.source}
\title{ddg.source sources a script & collects provenance}
\usage{
ddg.source(file, local = FALSE, echo = verbose, print.eval = echo,
  verbose = getOption("verbose"), max.deparse.length = 150, chdir = FALSE,
  encoding = getOption("encoding"), ignore.ddg.calls = TRUE)
}
\arguments{
\item{file}{- the name of the R script file to source.}

\item{local}{(optional) - the environment in which to evaluate parsed
expressions. If TRUE, the environment from which ddg.source is
called. If FALSE, the user's workspace (global environment).}

\item{echo}{(optional) - print each expression after parsing.}

\item{print.eval}{(optional) - print result of each evaluation.}

\item{verbose}{(optional) - print extra diagnostics.}

\item{max.deparse.length}{(optional) - maximum number of characters
output for deparse of a single expression.}

\item{chdir}{(optional) - change R working directory temporarily to
the directory containing the file to be sourced.}

\item{encoding}{(optional) - encoding to be assumed when file is a
character string.}

\item{ignore.ddg.calls}{(optional) - if TRUE, ignore DDG function calls.}

\item{ddg.source}{reads in an R script and executes it in the provided
enviroment. ddg.source essentially mimics the behaviour of the}
}
\value{
nothing
}
\description{
ddg.source sources a script & collects provenance
}
