% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{ddg.start}
\alias{ddg.start}
\title{ddg.start creates a procedure node of type Start called pname.
Users can right-click on a start node in DDG Explorer and see
the code between start and finish nodes in the original script.}
\usage{
ddg.start(pname = NULL)
}
\arguments{
\item{pname}{the label for the node.  This can be passed as
a string or as a name.}
}
\description{
ddg.start creates a procedure node of type Start called pname.
Users can right-click on a start node in DDG Explorer and see
the code between start and finish nodes in the original script.
}
