% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{prov.annotate.on}
\alias{prov.annotate.on}
\title{prov.annotate.on enables annotation for the specified functions. Functions not on
this list are not annotated. If fnames is NULL, all functions will be annotated.}
\usage{
prov.annotate.on(fnames = NULL)
}
\arguments{
\item{fnames}{- a list of one or more function names passed in as strings.}
}
\description{
prov.annotate.on enables annotation for the specified functions. Functions not on
this list are not annotated. If fnames is NULL, all functions will be annotated.
}
