% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{prov.clear.detail}
\alias{prov.clear.detail}
\title{prov.clear.detail clears the current value of provenance detail.}
\usage{
prov.clear.detail()
}
\value{
nothing
}
\description{
prov.clear.detail clears the current value of provenance detail.
}
