% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{prov.display}
\alias{prov.display}
\title{prov.display loads and displays the current provenance graph in DDG Explorer.}
\usage{
prov.display()
}
\value{
nothing
}
\description{
prov.display loads and displays the current provenance graph in DDG Explorer.
}
