% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{prov.get.detail}
\alias{prov.get.detail}
\title{prov.get.detail returns the current level of provenance detail.}
\usage{
prov.get.detail()
}
\value{
the current level of detail (0-3)
}
\description{
prov.get.detail returns the current level of provenance detail.
}
