% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{prov.init}
\alias{prov.init}
\title{prov.init intializes a new provenance graph. Called by the user
in console mode.}
\usage{
prov.init(r.script.path = NULL, prov.dir = NULL, overwrite = TRUE,
  annotate.inside.functions = FALSE, first.loop = 1, max.loops = 0,
  max.snapshot.size = 0, hash.algorithm = "md5")
}
\arguments{
\item{r.script.path}{the full path to the R script file
that is being executed. If provided, a copy of the script will
be saved with the provenance graph.}

\item{prov.dir}{the directory where the provenance graph will be 
saved. If not provided, the directory specified by the prov.dir 
option is used. Otherwise the R session temporary directory
is used.}

\item{overwrite}{if TRUE, generates a time stamp for the provenance
graph directory.}

\item{annotate.inside.functions}{if TRUE, provenance is collected 
inside functions.}

\item{first.loop}{the first loop to collect provenance in a for, 
while, or repeat statement.}

\item{max.loops}{the maximum number of loops to collect
provenance in a for, while, or repeat statement. If max.loops = -1,
there is no limit. If max.loops = 0, no loops are annotated. 
If non-zero, if-statements are also annotated.}

\item{max.snapshot.size}{the maximum size for objects that should
be output to snapshot files. If 0, no snapshot files are saved.
If -1, all snapshot files are saved. Size in kilobytes. Note that
this tests the size of the object that will be turned into a
snapshot, not the size of the resulting snapshot.}

\item{hash.algorithm}{the hash algorithm to use for files.
Choices are md5 (default), sha1, crc32, sha256, sha512, xxhash32, 
xxhash64 and murmur32. This feature uses the digest function from 
the digest package.}
}
\value{
nothing
}
\description{
prov.init intializes a new provenance graph. Called by the user
in console mode.
}
