% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loops.R
\name{prov.max.loops}
\alias{prov.max.loops}
\title{prov.max.loops returns the maximum number of iterations of a loop for
which provenance should be collected}
\usage{
prov.max.loops()
}
\value{
the maximum number of iterations to collect provenance
}
\description{
prov.max.loops returns the maximum number of iterations of a loop for
which provenance should be collected
}
