% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataNodes.R
\name{prov.max.snapshot.size}
\alias{prov.max.snapshot.size}
\title{prov.max.snapshot.size returns the maximum size for snapshots in KB}
\usage{
prov.max.snapshot.size()
}
\value{
maximum snapshot size in KB
}
\description{
prov.max.snapshot.size returns the maximum size for snapshots in KB
}
