% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_core.R
\name{.ddg.quit}
\alias{.ddg.quit}
\title{prov.auit}
\usage{
.ddg.quit(save.debug = FALSE)
}
\value{
prov.quit writes the current provenance to a file but does not 
  return a value.
}
\description{
prov.quit saves and closes the current provenance graph.
Called by the user in console mode.
}
