% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{prov.save}
\alias{prov.save}
\title{prov.save saves the current provenance graph.  If more R statements
are executed, they will be added to this ddg.  To finalize ddg,
call prov.quit, which will save and finalize the ddg.}
\usage{
prov.save(save.debug = FALSE)
}
\arguments{
\item{save.debug}{(optional) - If TRUE, save debug files to debug directory.}
}
\value{
nothing
}
\description{
prov.save saves the current provenance graph.  If more R statements
are executed, they will be added to this ddg.  To finalize ddg,
call prov.quit, which will save and finalize the ddg.
}
