% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{prov.set.detail}
\alias{prov.set.detail}
\title{prov.set.detail sets the level of detail for the provenance to be 
collected. If this value is not set, the level of detail is determined by the 
parameters of prov.run or prov.init.\cr
0 = no internal provenance, no snapshots.\cr
1 = 1 loop, snapshots < 10k.\cr
2 = 10 loops, snapshots < 100k.\cr
3 = all loops, all snapshots.}
\usage{
prov.set.detail(detail.level)
}
\arguments{
\item{detail.level}{level of detail to set (0-3)}
}
\value{
nothing
}
\description{
prov.set.detail sets the level of detail for the provenance to be 
collected. If this value is not set, the level of detail is determined by the 
parameters of prov.run or prov.init.\cr
0 = no internal provenance, no snapshots.\cr
1 = 1 loop, snapshots < 10k.\cr
2 = 10 loops, snapshots < 100k.\cr
3 = all loops, all snapshots.
}
