% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{prov.start}
\alias{prov.start}
\title{prov.start creates a procedure node of type Start called pname.
In DDG Explorer, the section of the provenance graph between
matching Start and Finish nodes can be expanded and collapsed. 
Users can also right-click on a Start or Finish node to see the 
intervening code in the original script.}
\usage{
prov.start(pname = NULL)
}
\arguments{
\item{pname}{the label for the node.  This can be passed as
a string or as a name.}
}
\value{
nothing
}
\description{
prov.start creates a procedure node of type Start called pname.
In DDG Explorer, the section of the provenance graph between
matching Start and Finish nodes can be expanded and collapsed. 
Users can also right-click on a Start or Finish node to see the 
intervening code in the original script.
}
