% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{ddg.save}
\alias{ddg.save}
\title{ddg.save saves the current provenance graph}
\usage{
ddg.save(save.debug = FALSE, quit = FALSE)
}
\arguments{
\item{save.debug}{(optional) - If TRUE, save debug files to debug directory.
Used in console mode.}

\item{quit}{(optional) - If TRUE, remove all DDG files from memory.}
}
\value{
nothing
}
\description{
ddg.save saves the current provenance graph
}
