% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{prov.annotate.off}
\alias{prov.annotate.off}
\title{prov.annotate.off disables annotation for the specified functions. Functions not on 
this list are annotated. If fnames is NULL, no functions will be annotated.}
\usage{
prov.annotate.off(fnames = NULL)
}
\arguments{
\item{fnames}{a list of one or more function names passed in as strings.}
}
\description{
prov.annotate.off disables annotation for the specified functions. Functions not on 
this list are annotated. If fnames is NULL, no functions will be annotated.
}
