% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{prov.json}
\alias{prov.json}
\title{prov.json returns the current provenance graph as a prov-json string}
\usage{
prov.json()
}
\value{
the current provenance graph
}
\description{
prov.json returns the current provenance graph as a prov-json string
}
