% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{prov.quit}
\alias{prov.quit}
\title{prov.quit saves and closes the current provenance graph.}
\usage{
prov.quit(save.debug = FALSE)
}
\arguments{
\item{save.debug}{(optional) - If TRUE, save debug files to debug directory.
Used in console mode.}
}
\value{
nothing
}
\description{
prov.quit saves and closes the current provenance graph.
}
