% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{prov.run}
\alias{prov.run}
\title{prov.run initiates execution of a script and collects provenance as 
the script executes.}
\usage{
prov.run(r.script.path = NULL, prov.dir = NULL, overwrite = TRUE,
  f = NULL, annotate.inside.functions = FALSE, first.loop = 1,
  max.loops = 0, max.snapshot.size = 0, save.debug = FALSE,
  display = FALSE, hash.algorithm = "md5")
}
\arguments{
\item{r.script.path}{the full path to the R script file
that is being executed. If provided, a copy of the script will
be saved with the provenance graph.}

\item{prov.dir}{the directory where the provenance graph will be saved. 
If not provided, the directory specified by the prov.dir option is used. 
Otherwise the R session temporary directory is used.}

\item{overwrite}{if TRUE, generates a time stamp for the provenance 
graph directory.}

\item{f}{a function to run. If supplied, the function f is executed 
with calls to prov.init and prov.save so that provenance for the 
function is captured.  Exactly one of f and r.script.path should be provided.}

\item{annotate.inside.functions}{if TRUE, provenance is collected 
inside functions.}

\item{first.loop}{the first loop to collect provenance in a for, 
while, or repeat statement.}

\item{max.loops}{the maximum number of loops to collect provenance 
in a for, while, or repeat statement. If max.loops = -1,
there is no limit. If max.loops = 0, no loops are annotated. 
If non-zero, if-statements are also annotated.}

\item{max.snapshot.size}{the maximum size for objects that should 
be output to snapshot files. If 0, no snapshot files are saved.
If -1, all snapshot files are saved. Size in kilobytes. Note that
this tests the size of the object that will be turned into a
snapshot, not the size of the resulting snapshot.}

\item{save.debug}{if TRUE, debug files are saved to the debug 
directory.}

\item{display}{if TRUE, the provenance graph is displayed in DDG Explorer}

\item{hash.algorithm}{the hash algorithm to use for files.
Choices are md5 (default), sha1, crc32, sha256, sha512, xxhash32, 
xxhash64 and murmur32. This feature uses the digest function from 
the digest package.}
}
\value{
nothing
}
\description{
prov.run initiates execution of a script and collects provenance as 
the script executes.
}
