% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{prov.save}
\alias{prov.save}
\title{prov.save saves the current provenance graph to a prov-json file.
If more R statements are executed, the provenance for these statements
is added to the graph. The graph is finalized with prov.quit.
Called by the user in console mode.}
\usage{
prov.save(save.debug = FALSE)
}
\arguments{
\item{save.debug}{If TRUE, debug files are saved to the debug directory}
}
\value{
nothing
}
\description{
prov.save saves the current provenance graph to a prov-json file.
If more R statements are executed, the provenance for these statements
is added to the graph. The graph is finalized with prov.quit.
Called by the user in console mode.
}
