% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{prov.source}
\alias{prov.source}
\title{prov.source reads in an R script and executes it in the provided
enviroment. prov.source essentially mimics the behaviour of the}
\usage{
prov.source(file, local = FALSE, echo = verbose, print.eval = echo,
  verbose = getOption("verbose"), max.deparse.length = 150, chdir = FALSE,
  encoding = getOption("encoding"), ignore.ddg.calls = TRUE,
  calling.script = NA, startLine = NA, startCol = NA, endLine = NA,
  endCol = NA)
}
\arguments{
\item{file}{- the name of the R script file to source.}

\item{local}{(optional) - the environment in which to evaluate parsed
expressions. If TRUE, the environment from which prov.source is
called. If FALSE, the user's workspace (global environment).}

\item{echo}{(optional) - print each expression after parsing.}

\item{print.eval}{(optional) - print result of each evaluation.}

\item{verbose}{(optional) - print extra diagnostics.}

\item{max.deparse.length}{(optional) - maximum number of characters
output for deparse of a single expression.}

\item{chdir}{(optional) - change R working directory temporarily to
the directory containing the file to be sourced.}

\item{encoding}{(optional) - encoding to be assumed when file is a
character string.}

\item{ignore.ddg.calls}{(optional) - if TRUE, ignore DDG function calls.}

\item{calling.script}{(optional) - the number of the script that source was called in}

\item{startLine}{(optional) - the line that the source call starts on}

\item{startCol}{(optional) - the column that the source call starts on}

\item{endLine}{(optional) - the line that the source call ends on}

\item{endCol}{(optional) - the column that the source call ends on}
}
\value{
nothing
}
\description{
prov.source reads in an R script and executes it in the provided
enviroment. prov.source essentially mimics the behaviour of the
}
