\name{RDataTracker-package}
\title{RDataTracker Data Provenance Collector}
\alias{RDataTracker}
\docType{package}
\description{RDataTracker is a collection of R functions that allows a user to 
annotate (instrument) an R script so that data provenance in the 
form of a Data Derivation Graph (DDG) is created whenever the script 
executes. The DDG can then be viewed, queried, and stored using a 
separate tool called DDG Explorer.}
